requirements.txt
grpcio
grpcio-tools
grpcio-reflection
confluent-kafka
minio